numcodecs_jpeg2000

Function decompress

pub fn decompress(encoded: &[u8]) -> Result<AnyArray, Jpeg2000CodecError>

Decompress the encoded data into an array using JPEG 2000.

Errors

Errors with

Jpeg2000CodecError::HeaderDecodeFailed if decoding the header failed

Jpeg2000CodecError::SliceDecodeFailed if decoding a slice failed

Jpeg2000CodecError::Jpeg2000DecodeFailed if decoding with JPEG 2000 failed

Jpeg2000CodecError::DecodeInvalidShape if the encoded data decodes to an unexpected shape

Jpeg2000CodecError::DecodeTooManySlices if the encoded data contains too many slices